requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
docx2txt==0.8
emoji==2.14.0
eval_type_backport==0.2.0
fastapi==0.115.6
filetype==1.2.0
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
isort==5.13.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.24
langchain-ollama==0.2.1
langchain-openai==0.2.12
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.2.2
lxml==5.3.0
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.4.4
openai==1.57.2
orjson==3.10.12
packaging==24.2
pathspec==0.12.1
pdfminer==20191125
pdfminer.six==20240706
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
propcache==0.2.1
psutil==6.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.19
python-oxmsg==0.0.1
python-pptx==1.0.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.16.11
unstructured-client==0.28.1
urllib3==2.2.3
uvicorn==0.32.1
webencodings==0.5.1
wrapt==1.17.0
XlsxWriter==3.2.0
yarl==1.18.3
